feat: create Teacher Menu DB and API
레스너 찾기 DB 및 API 추가

+++ clnewze-back/src/main/java/com/clnewze/back/clnewzeback/domain/TeacherMenuCategory.java
@@ -4,8 +4,8 @@ import lombok.Data;
 
 @Data
 public class TeacherMenuCategory {
-  int mtno;
-  int orderby;
+  Long mtno;
+  Long orderby;
   String name;
   String category;
 }
